back-end/requirements.txt
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==3.3.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-yasg==1.21.7
eskiz-sms==0.2.3
et-xmlfile==1.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
psycopg==3.1.17
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rfc3986==1.5.0
rpds-py==0.17.1
sniffio==1.3.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
xlrd==2.0.1
xlwt==1.3.0
